requirements.txt
discord.py
python-dotenv
huggingface_hub
langchain-community
langchain_text_splitters
chromadb
transformers

